+++ .github/workflows/ci.yml
@@ -93,4 +93,3 @@ jobs:
       - run: npx playwright install --with-deps chromium
       - run: npm run build
       - run: npm run test:e2e
-        continue-on-error: true
